+++ app/modules/editor/redux/actions/composed/changeEnduranceLaps.ts
@@ -1,6 +1,7 @@
+import Immutable from 'immutable';
 import { StoreThunkAction } from '../../../../../redux/types';
 import { TableTypeEnum } from '../../../enums';
-import createTable from '../../../pure/createTable';
+import createTable from '../../../helpers/createTable';
 import setEditorInitialStateAction from '../setEditorInitialStateAction';
 
 export type ChangeEnduranceLapsType = (amount: number) => StoreThunkAction;
@@ -10,5 +11,5 @@ export const changeEnduranceLaps: ChangeEnduranceLapsType = laps => (dispatch, g
   const base = editor.getIn<number>(['trainingTable', 'base']);
   const baseBreaks = editor.getIn<number>(['trainingTable', 'baseBreaks']);
   const newState = createTable(base, baseBreaks, TableTypeEnum.TABLE_TYPE_ENDURANCE, laps);
-  return dispatch(setEditorInitialStateAction(newState));
+  return dispatch(setEditorInitialStateAction(Immutable.fromJS(newState)));
 };
